Add compile option to use skim binary instead of fzf binary
