+++ Cargo.toml
@@ -14,6 +14,10 @@ license = "Apache-2.0"
 [badges]
 travis-ci = { repository = "denisidoro/navi", branch = "master" }
 
+[features]
+default = []
+use_skim = []
+
 [dependencies]
 regex = "1.3.4"
 structopt = "0.3"
