+++ src/structures/mod.rs
@@ -1,5 +1,5 @@
 pub mod cheat;
 pub mod error;
 pub mod fnv;
-pub mod fzf;
+pub mod finder;
 pub mod option;
