+++ src/structures/option.rs
@@ -37,13 +37,13 @@ pub struct Config {
     #[structopt(long)]
     pub no_preview: bool,
 
-    /// FZF overrides for cheat selection
-    #[structopt(long, env = "NAVI_FZF_OVERRIDES")]
-    pub fzf_overrides: Option<String>,
+    /// finder overrides for cheat selection
+    #[structopt(long, env = "NAVI_FINDER_OVERRIDES")]
+    pub finder_overrides: Option<String>,
 
-    /// FZF overrides for variable selection
-    #[structopt(long, env = "NAVI_FZF_OVERRIDES_VAR")]
-    pub fzf_overrides_var: Option<String>,
+    /// finder overrides for variable selection
+    #[structopt(long, env = "NAVI_FINDER_OVERRIDES_VAR")]
+    pub finder_overrides_var: Option<String>,
 
     #[structopt(subcommand)]
     pub cmd: Option<Command>,
